Move download functionality into download
